Add support for different tabs in channels
