+++ extractor/src/main/java/org/schabi/newpipe/extractor/channel/ChannelTabExtractor.java
@@ -1,0 +1,12 @@
+package org.schabi.newpipe.extractor.channel;
+
+import org.schabi.newpipe.extractor.InfoItem;
+import org.schabi.newpipe.extractor.ListExtractor;
+import org.schabi.newpipe.extractor.StreamingService;
+import org.schabi.newpipe.extractor.linkhandler.ListLinkHandler;
+
+public abstract class ChannelTabExtractor<R extends InfoItem>  extends ListExtractor<R> {
+    public ChannelTabExtractor(StreamingService service, ListLinkHandler linkHandler) {
+        super(service, linkHandler);
+    }
+}
