+++ extractor/src/main/java/org/schabi/newpipe/extractor/services/media_ccc/extractors/MediaCCCConferenceEventsExtractor.java
@@ -1,0 +1,58 @@
+package org.schabi.newpipe.extractor.services.media_ccc.extractors;
+
+import com.grack.nanojson.JsonArray;
+import com.grack.nanojson.JsonObject;
+
+import org.schabi.newpipe.extractor.StreamingService;
+import org.schabi.newpipe.extractor.channel.ChannelStreamsExtractor;
+import org.schabi.newpipe.extractor.downloader.Downloader;
+import org.schabi.newpipe.extractor.exceptions.ExtractionException;
+import org.schabi.newpipe.extractor.exceptions.ParsingException;
+import org.schabi.newpipe.extractor.linkhandler.ListLinkHandler;
+import org.schabi.newpipe.extractor.services.media_ccc.extractors.infoItems.MediaCCCStreamInfoItemExtractor;
+import org.schabi.newpipe.extractor.stream.StreamInfoItem;
+import org.schabi.newpipe.extractor.stream.StreamInfoItemsCollector;
+
+import java.io.IOException;
+
+import javax.annotation.Nonnull;
+
+public class MediaCCCConferenceEventsExtractor extends ChannelStreamsExtractor {
+    private JsonObject conferenceData;
+
+    public MediaCCCConferenceEventsExtractor(StreamingService service, ListLinkHandler linkHandler, JsonObject conferenceData) {
+        super(service, linkHandler);
+
+        this.conferenceData = conferenceData;
+    }
+
+    @Nonnull
+    @Override
+    public InfoItemsPage<StreamInfoItem> getInitialPage() throws IOException, ExtractionException {
+        StreamInfoItemsCollector collector = new StreamInfoItemsCollector(getServiceId());
+        JsonArray events = conferenceData.getArray("events");
+        for (int i = 0; i < events.size(); i++) {
+            collector.commit(new MediaCCCStreamInfoItemExtractor(events.getObject(i)));
+        }
+        return new InfoItemsPage<>(collector, null);
+    }
+
+    @Override
+    public String getNextPageUrl() throws IOException, ExtractionException {
+        return null;
+    }
+
+    @Override
+    public InfoItemsPage<StreamInfoItem> getPage(String pageUrl) throws IOException, ExtractionException {
+        return null;
+    }
+
+    @Override
+    public void onFetchPage(@Nonnull Downloader downloader) throws IOException, ExtractionException {}
+
+    @Nonnull
+    @Override
+    public String getName() throws ParsingException {
+        return "Events";
+    }
+}
