+++ extractor/src/main/java/org/schabi/newpipe/extractor/services/media_ccc/extractors/MediaCCCConferenceExtractor.java
@@ -1,24 +1,24 @@
 package org.schabi.newpipe.extractor.services.media_ccc.extractors;
 
-import com.grack.nanojson.JsonArray;
 import com.grack.nanojson.JsonObject;
 import com.grack.nanojson.JsonParser;
 import com.grack.nanojson.JsonParserException;
+
 import org.schabi.newpipe.extractor.StreamingService;
 import org.schabi.newpipe.extractor.channel.ChannelExtractor;
+import org.schabi.newpipe.extractor.channel.ChannelTabExtractor;
 import org.schabi.newpipe.extractor.downloader.Downloader;
 import org.schabi.newpipe.extractor.exceptions.ExtractionException;
 import org.schabi.newpipe.extractor.exceptions.ParsingException;
 import org.schabi.newpipe.extractor.linkhandler.ListLinkHandler;
-import org.schabi.newpipe.extractor.services.media_ccc.extractors.infoItems.MediaCCCStreamInfoItemExtractor;
-import org.schabi.newpipe.extractor.stream.StreamInfoItem;
-import org.schabi.newpipe.extractor.stream.StreamInfoItemsCollector;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 import javax.annotation.Nonnull;
-import java.io.IOException;
 
 public class MediaCCCConferenceExtractor extends ChannelExtractor {
-
     private JsonObject conferenceData;
 
     public MediaCCCConferenceExtractor(StreamingService service, ListLinkHandler linkHandler) {
@@ -50,25 +50,13 @@ public class MediaCCCConferenceExtractor extends ChannelExtractor {
         return null;
     }
 
-    @Nonnull
     @Override
-    public InfoItemsPage<StreamInfoItem> getInitialPage() throws IOException, ExtractionException {
-        StreamInfoItemsCollector collector = new StreamInfoItemsCollector(getServiceId());
-        JsonArray events = conferenceData.getArray("events");
-        for (int i = 0; i < events.size(); i++) {
-            collector.commit(new MediaCCCStreamInfoItemExtractor(events.getObject(i)));
-        }
-        return new InfoItemsPage<>(collector, null);
-    }
+    public List<ChannelTabExtractor> getTabs() throws ParsingException {
+        List<ChannelTabExtractor> tabs = new ArrayList<>();
 
-    @Override
-    public String getNextPageUrl() throws IOException, ExtractionException {
-        return null;
-    }
+        tabs.add(new MediaCCCConferenceEventsExtractor(getService(), (ListLinkHandler) getLinkHandler(), conferenceData));
 
-    @Override
-    public InfoItemsPage<StreamInfoItem> getPage(String pageUrl) throws IOException, ExtractionException {
-        return null;
+        return tabs;
     }
 
     @Override
