+++ extractor/src/main/java/org/schabi/newpipe/extractor/services/soundcloud/SoundcloudChannelExtractor.java
@@ -4,25 +4,25 @@ import com.grack.nanojson.JsonArray;
 import com.grack.nanojson.JsonObject;
 import com.grack.nanojson.JsonParser;
 import com.grack.nanojson.JsonParserException;
+
 import org.schabi.newpipe.extractor.StreamingService;
 import org.schabi.newpipe.extractor.channel.ChannelExtractor;
+import org.schabi.newpipe.extractor.channel.ChannelTabExtractor;
 import org.schabi.newpipe.extractor.downloader.Downloader;
 import org.schabi.newpipe.extractor.exceptions.ExtractionException;
 import org.schabi.newpipe.extractor.exceptions.ParsingException;
 import org.schabi.newpipe.extractor.linkhandler.ListLinkHandler;
-import org.schabi.newpipe.extractor.stream.StreamInfoItem;
-import org.schabi.newpipe.extractor.stream.StreamInfoItemsCollector;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 import javax.annotation.Nonnull;
-import java.io.IOException;
 
 @SuppressWarnings("WeakerAccess")
 public class SoundcloudChannelExtractor extends ChannelExtractor {
     private String userId;
     private JsonObject user;
-
-    private StreamInfoItemsCollector streamInfoItemsCollector = null;
-    private String nextPageUrl = null;
 
     public SoundcloudChannelExtractor(StreamingService service, ListLinkHandler linkHandler) {
         super(service, linkHandler);
@@ -83,47 +83,12 @@ public class SoundcloudChannelExtractor extends ChannelExtractor {
         return user.getString("description", "");
     }
 
-    @Nonnull
     @Override
-    public InfoItemsPage<StreamInfoItem> getInitialPage() throws ExtractionException {
-        if (streamInfoItemsCollector == null) {
-            computeNextPageAndGetStreams();
-        }
-        return new InfoItemsPage<>(streamInfoItemsCollector, getNextPageUrl());
-    }
+    public List<ChannelTabExtractor> getTabs() {
+        List<ChannelTabExtractor> tabs = new ArrayList<>();
 
-    @Override
-    public String getNextPageUrl() throws ExtractionException {
-        if (nextPageUrl == null) {
-            computeNextPageAndGetStreams();
-        }
-        return nextPageUrl;
-    }
+        tabs.add(new SoundcloudChannelTracksExtractor(getService(), (ListLinkHandler) getLinkHandler()));
 
-    private void computeNextPageAndGetStreams() throws ExtractionException {
-        try {
-            streamInfoItemsCollector = new StreamInfoItemsCollector(getServiceId());
-
-            String apiUrl = "https://api-v2.soundcloud.com/users/" + getId() + "/tracks"
-                    + "?client_id=" + SoundcloudParsingHelper.clientId()
-                    + "&limit=20"
-                    + "&linked_partitioning=1";
-
-            nextPageUrl = SoundcloudParsingHelper.getStreamsFromApiMinItems(15, streamInfoItemsCollector, apiUrl);
-        } catch (Exception e) {
-            throw new ExtractionException("Could not get next page", e);
-        }
-    }
-
-    @Override
-    public InfoItemsPage<StreamInfoItem> getPage(final String pageUrl) throws IOException, ExtractionException {
-        if (pageUrl == null || pageUrl.isEmpty()) {
-            throw new ExtractionException(new IllegalArgumentException("Page url is empty or null"));
-        }
-
-        StreamInfoItemsCollector collector = new StreamInfoItemsCollector(getServiceId());
-        String nextPageUrl = SoundcloudParsingHelper.getStreamsFromApiMinItems(15, collector, pageUrl);
-
-        return new InfoItemsPage<>(collector, nextPageUrl);
+        return tabs;
     }
 }
