+++ extractor/src/main/java/org/schabi/newpipe/extractor/services/youtube/extractors/YoutubeChannelVideosExtractor.java
@@ -1,0 +1,115 @@
+package org.schabi.newpipe.extractor.services.youtube.extractors;
+
+import com.grack.nanojson.JsonArray;
+import com.grack.nanojson.JsonObject;
+
+import org.schabi.newpipe.extractor.StreamingService;
+import org.schabi.newpipe.extractor.channel.ChannelStreamsExtractor;
+import org.schabi.newpipe.extractor.downloader.Downloader;
+import org.schabi.newpipe.extractor.exceptions.ExtractionException;
+import org.schabi.newpipe.extractor.exceptions.ParsingException;
+import org.schabi.newpipe.extractor.linkhandler.ListLinkHandler;
+import org.schabi.newpipe.extractor.localization.TimeAgoParser;
+import org.schabi.newpipe.extractor.stream.StreamInfoItem;
+import org.schabi.newpipe.extractor.stream.StreamInfoItemsCollector;
+
+import java.io.IOException;
+
+import javax.annotation.Nonnull;
+
+import static org.schabi.newpipe.extractor.services.youtube.linkHandler.YoutubeParsingHelper.getJsonResponse;
+
+public class YoutubeChannelVideosExtractor extends ChannelStreamsExtractor {
+    private JsonObject videoTab;
+
+    public YoutubeChannelVideosExtractor(StreamingService service, ListLinkHandler linkHandler, JsonObject videoTab) {
+        super(service, linkHandler);
+
+        this.videoTab = videoTab;
+    }
+
+    @Override
+    public void onFetchPage(@Nonnull Downloader downloader) throws IOException, ExtractionException {}
+
+    @Override
+    public String getNextPageUrl() throws ExtractionException {
+        return getNextPageUrlFrom(videoTab.getObject("content").getObject("sectionListRenderer")
+                .getArray("contents").getObject(0).getObject("itemSectionRenderer")
+                .getArray("contents").getObject(0).getObject("gridRenderer").getArray("continuations"));
+    }
+    @Nonnull
+    @Override
+    public String getName() throws ParsingException {
+        return "Videos";
+    }
+
+    @Nonnull
+    @Override
+    public InfoItemsPage<StreamInfoItem> getInitialPage() throws ExtractionException {
+        StreamInfoItemsCollector collector = new StreamInfoItemsCollector(getServiceId());
+
+        JsonArray videos = videoTab.getObject("content").getObject("sectionListRenderer").getArray("contents")
+                .getObject(0).getObject("itemSectionRenderer").getArray("contents").getObject(0)
+                .getObject("gridRenderer").getArray("items");
+        collectStreamsFrom(collector, videos);
+
+        return new InfoItemsPage<>(collector, getNextPageUrl());
+    }
+
+    @Override
+    public InfoItemsPage<StreamInfoItem> getPage(String pageUrl) throws IOException, ExtractionException {
+        if (pageUrl == null || pageUrl.isEmpty()) {
+            throw new ExtractionException(new IllegalArgumentException("Page url is empty or null"));
+        }
+
+        // Unfortunately, we have to fetch the page even if we are only getting next streams,
+        // as they don't deliver enough information on their own (the channel name, for example).
+        fetchPage();
+
+        StreamInfoItemsCollector collector = new StreamInfoItemsCollector(getServiceId());
+        final JsonArray ajaxJson = getJsonResponse(pageUrl, getExtractorLocalization());
+
+        JsonObject sectionListContinuation = ajaxJson.getObject(1).getObject("response")
+                .getObject("continuationContents").getObject("gridContinuation");
+
+        collectStreamsFrom(collector, sectionListContinuation.getArray("items"));
+
+        return new InfoItemsPage<>(collector, getNextPageUrlFrom(sectionListContinuation.getArray("continuations")));
+    }
+
+
+    private String getNextPageUrlFrom(JsonArray continuations) {
+        if (continuations == null) return "";
+
+        JsonObject nextContinuationData = continuations.getObject(0).getObject("nextContinuationData");
+        String continuation = nextContinuationData.getString("continuation");
+        String clickTrackingParams = nextContinuationData.getString("clickTrackingParams");
+        return "https://www.youtube.com/browse_ajax?ctoken=" + continuation + "&continuation=" + continuation
+                + "&itct=" + clickTrackingParams;
+    }
+
+    private void collectStreamsFrom(StreamInfoItemsCollector collector, JsonArray videos) throws ParsingException {
+        collector.reset();
+
+        final String uploaderName = getName();
+        final String uploaderUrl = getUrl();
+        final TimeAgoParser timeAgoParser = getTimeAgoParser();
+
+        for (Object video : videos) {
+            if (((JsonObject) video).getObject("gridVideoRenderer") != null) {
+                collector.commit(new YoutubeStreamInfoItemExtractor(
+                        ((JsonObject) video).getObject("gridVideoRenderer"), timeAgoParser) {
+                    @Override
+                    public String getUploaderName() {
+                        return uploaderName;
+                    }
+
+                    @Override
+                    public String getUploaderUrl() {
+                        return uploaderUrl;
+                    }
+                });
+            }
+        }
+    }
+}
