+++ extractor/src/test/java/org/schabi/newpipe/extractor/services/youtube/YoutubeChannelExtractorTest.java
@@ -5,15 +5,20 @@ import org.junit.Test;
 import org.schabi.newpipe.DownloaderTestImpl;
 import org.schabi.newpipe.extractor.NewPipe;
 import org.schabi.newpipe.extractor.channel.ChannelExtractor;
+import org.schabi.newpipe.extractor.channel.ChannelTabExtractor;
 import org.schabi.newpipe.extractor.exceptions.ContentNotAvailableException;
 import org.schabi.newpipe.extractor.exceptions.ParsingException;
 import org.schabi.newpipe.extractor.services.BaseChannelExtractorTest;
 import org.schabi.newpipe.extractor.services.youtube.extractors.YoutubeChannelExtractor;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import static org.schabi.newpipe.extractor.ExtractorAsserts.assertIsSecureUrl;
 import static org.schabi.newpipe.extractor.ServiceList.YouTube;
-import static org.schabi.newpipe.extractor.services.DefaultTests.*;
+import static org.schabi.newpipe.extractor.services.DefaultTests.defaultTestGetPageInNewExtractor;
+import static org.schabi.newpipe.extractor.services.DefaultTests.defaultTestMoreItems;
+import static org.schabi.newpipe.extractor.services.DefaultTests.defaultTestRelatedItems;
 
 /**
  * Test for {@link ChannelExtractor}
@@ -87,12 +92,16 @@ public class YoutubeChannelExtractorTest {
 
         @Test
         public void testRelatedItems() throws Exception {
-            defaultTestRelatedItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestRelatedItems(tab);
+            }
         }
 
         @Test
         public void testMoreRelatedItems() throws Exception {
-            defaultTestMoreItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestMoreItems(tab);
+            }
         }
 
          /*//////////////////////////////////////////////////////////////////////////
@@ -177,12 +186,16 @@ public class YoutubeChannelExtractorTest {
 
         @Test
         public void testRelatedItems() throws Exception {
-            defaultTestRelatedItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestRelatedItems(tab);
+            }
         }
 
         @Test
         public void testMoreRelatedItems() throws Exception {
-            defaultTestMoreItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestMoreItems(tab);
+            }
         }
 
          /*//////////////////////////////////////////////////////////////////////////
@@ -239,8 +252,11 @@ public class YoutubeChannelExtractorTest {
 
         @Test
         public void testGetPageInNewExtractor() throws Exception {
-            final ChannelExtractor newExtractor = YouTube.getChannelExtractor(extractor.getUrl());
-            defaultTestGetPageInNewExtractor(extractor, newExtractor);
+            for (int i = 0; i < extractor.getTabs().size(); i++) {
+                final ChannelExtractor newExtractor = YouTube.getChannelExtractor(extractor.getUrl());
+                newExtractor.fetchPage();
+                defaultTestGetPageInNewExtractor(extractor.getTabs().get(i), newExtractor.getTabs().get(i));
+            }
         }
 
         /*//////////////////////////////////////////////////////////////////////////
@@ -279,12 +295,16 @@ public class YoutubeChannelExtractorTest {
 
         @Test
         public void testRelatedItems() throws Exception {
-            defaultTestRelatedItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestRelatedItems(tab);
+            }
         }
 
         @Test
         public void testMoreRelatedItems() throws Exception {
-            defaultTestMoreItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestMoreItems(tab);
+            }
         }
 
          /*//////////////////////////////////////////////////////////////////////////
@@ -370,12 +390,16 @@ public class YoutubeChannelExtractorTest {
 
         @Test
         public void testRelatedItems() throws Exception {
-            defaultTestRelatedItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestRelatedItems(tab);
+            }
         }
 
         @Test
         public void testMoreRelatedItems() throws Exception {
-            defaultTestMoreItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestMoreItems(tab);
+            }
         }
 
          /*//////////////////////////////////////////////////////////////////////////
@@ -460,12 +484,16 @@ public class YoutubeChannelExtractorTest {
 
         @Test
         public void testRelatedItems() throws Exception {
-            defaultTestRelatedItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestRelatedItems(tab);
+            }
         }
 
         @Test
         public void testMoreRelatedItems() throws Exception {
-            defaultTestMoreItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestMoreItems(tab);
+            }
         }
 
          /*//////////////////////////////////////////////////////////////////////////
@@ -552,13 +580,17 @@ public class YoutubeChannelExtractorTest {
 
         @Test
         public void testRelatedItems() throws Exception {
-            defaultTestRelatedItems(extractor);
+            for (ChannelTabExtractor tab : extractor.getTabs()) {
+                defaultTestRelatedItems(tab);
+            }
         }
 
         @Test
         public void testMoreRelatedItems() {
             try {
-                defaultTestMoreItems(extractor);
+                for (ChannelTabExtractor tab : extractor.getTabs()) {
+                    defaultTestMoreItems(tab);
+                }
             } catch (Throwable ignored) {
                 return;
             }
